finished reaper, huge load, and bombs away conversion

+++ Assets/Game/Scripts/RulesetScripts/Events/Reaper/TheReaperComes.cs
@@ -17,46 +17,48 @@ public class TheReaperComes : GameEvent
 
     private void InitReapers()
     {
-		PlayerManager[] players = PlayerWrangler.GetAllPlayers();
+        PlayerManager[] players = PlayerWrangler.GetAllPlayers();
 
-		for (byte i = 0; i < players.Length; i++)
+        for (byte i = 0; i < players.Length; i++)
         {
-			int viewID = PhotonNetwork.AllocateViewID ();
-			PhotonView.RPC ("RPC_SpawnReapers", PhotonTargets.All, viewID, (byte) i, players[i].name);
+            int viewID = PhotonNetwork.AllocateViewID();
+            PhotonView.RPC("RPC_SpawnReapers", PhotonTargets.All, viewID, i, players[i].name);
         }
     }
 
-	[PunRPC]
-	void RPC_SpawnReapers(int _viewID, byte index, string targetID)
-	{
-		GameObject newReaper = Instantiate (reaper);
-		Reaper _reaper = newReaper.GetComponent<Reaper> ();
-		newReaper.GetComponent<PhotonView> ().viewID = _viewID;
-		reapers.Add(_reaper);
-		_reaper.enabled = true;
-		_reaper.SetTargetPlayer (PlayerWrangler.GetPlayer (targetID));
-		_reaper.SetSpawnPoint (reaperSpawns [index]);
-		_reaper.SetPoints (pointsPlayerLosesOnDeath);
-		_reaper.Setup ();
+    [PunRPC]
+    void RPC_SpawnReapers(int _viewID, byte index, string targetID)
+    {
+        GameObject newReaper = Instantiate(reaper);
+        Reaper _reaper = newReaper.GetComponent<Reaper>();
+        newReaper.GetComponent<PhotonView>().viewID = _viewID;
+        reapers.Add(_reaper);
+        _reaper.enabled = true;
+        _reaper.SetTargetPlayer(PlayerWrangler.GetPlayer(targetID));
+        _reaper.SetSpawnPoint(reaperSpawns[index]);
+        _reaper.SetPoints(pointsPlayerLosesOnDeath);
+        _reaper.Setup();
 
-	}
+    }
 
     public override void StartEvent()
     {
         foreach (GameObject go in objectsToSetActive)
             go.SetActive(false);
 
-		if (PhotonNetwork.isMasterClient && reapers.Count < 1)
+        if (PhotonNetwork.isMasterClient && reapers.Count < 1)
             InitReapers();
 
-		byte num = PlayerWrangler.GetNumOfPlayers();
+        byte num = PlayerWrangler.GetNumOfPlayers();
 
-		if (reapers.Count > 0) {
-			for (byte index = 0; index < num; index++) {
-				reapers [index].enabled = true;
-				reapers [index].Setup ();
-			}
-		}
+        if (reapers.Count > 0)
+        {
+            for (byte index = 0; index < num; index++)
+            {
+                reapers[index].enabled = true;
+                reapers[index].Setup();
+            }
+        }
 
         gameEventDur = StartCoroutine(EventDuration());
     }
